ml-service/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.10.6
python-dotenv==1.0.1
numpy==2.0.2
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.1
tensorflow==2.18.0
